pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "lxml>=5.4.0",
    "python-telegram-bot>=20.0,<21.0",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
